refactor(desktop): split timeline message rows
